Add rsync option to iotstack_backup
Adds rsync option as an alternative to scp. With scp, backup files are
simply copied to the target host. There is no concept of pruning the
destination folders which will simply grow over time until the user
takes some other action. Switching to rsync means that the auto-pruning
that occurs on the RPi is mirrored on the target host.

Documentation enhanced to descrive rsync option. Table of contents
added to documentation to make it easier to navigate.