Documentation improvements
Clarify requirement to match restore-side with backup-side if the user
chooses "roll your own" or "invert the problem" methods for preserving
backups off the RPi.

Change bare metal restore instructions to mimic the menu rather than
run the menu (which ensures up-to-date docker-compose and minimises
risk of doing things in the wrong order).

Consistent references to .local/bin.