Adds RCLONE support
This version adds support for RCLONE. It does this by removing the
need to alter environment variables in both `iotstack_backup` and
`iotstack_restore` in favour of a configuration file at:

```
~/.config/iotstack_backup/config.yml
```

The config file is where you specify both a backup and restore method.
The choices are SCP, RSYNC and RCLONE.

There is extensive documentation for getting RCLONE set up as a
Dropbox client.