Support for volumes being on separate media
Supports the possibility of `~/IOTstack/volumes` being a mount-point so that the storage is on separate media:

1. Adds an exclusion for `lost+found` to `iotstack_backup_general`.
2. Changes `iotstack_restore_general` to iterate the **restored** `services` and `volumes` directories, removing the corresponding file system component of the **existing** `services` and `volumes` directories before moving the **restored** file system component into place.

The second change is necessary if `~/IOTstack/volumes` is a mount-point because it prevents the attempt to remove it.

The change also has the (beneficial) side-effect of making a restore behave like a "merge". In other words, if a container exists on the system being restored but is not represented in the backup, its components (`services` and `volumes`) will remain unchanged.

> Previously, the whole of `services` and `volumes` were removed, unconditionally, likely leading to data loss.