consequential changes from supporting hass.io in PiBuilder
Adding hass.io installation as an option in PiBuilder exposed weaknesses
in how IOTstackBackup checked whether IOTstack (as a whole) or a
container installed under IOTstack was running.

It boils down to the difference between `docker ps` and
`docker-compose ps`.

This change prevents IOTstackBackup and, in particular, the "bare metal
restore" scenario from mis-reading the presence of containers managed
by `hassio_supervisor` as indicating "some part of IOTstack is running".