fix bug when including environment files
Existing code attempts to add `.env` and `*.env` in for-loop expansion,
assuming the loop will be a no-op if no such files exist. What actually
happens is:

1. `.env` is always specified as an inclusion.
2. If no files matching `*.env` exist then `*.env` is specified as an
inclusion.

Each of those causes `tar` to object and exit with an error. This does
not damage the final `.tar.gz` and doesn't affect any subsequent
restores but it's untidy.

Signed-off-by: Phill Kelley <34226495+Paraphraser@users.noreply.github.com>