support passing arbitrary options to scp, rsync and rclone
[Discord question](https://discord.com/channels/638610460567928832/638610461109256194/1136669857052631061)
reported a problem apparently triggered by the transition from the
Secure Copy Protocol (SCP) to the Secure File Transfer Protocol (SFTP).
The question noted the problem could be avoided by passing the `-O`
flag to the `scp` command, which forces `scp` to use SCP rather than
SFTP. The question asked whether this could be supported as a formal
option?

This feature adds support for an optional `options:` clause in the
configuration YAML. It is supported, independently, for both the backup
and restore groupings. If present and non-empty, the associated value
will be passed to whichever principal command is being invoked (scp,
rsync or rclone) by the backup/restore method.

A commented-out `options: "-O"` has been added to the relevant
configuration templates. The clauses are disabled by default because
the `-O` flag is not supported by all versions of `scp` and causes a
crash if it is passed when not supported. The feature must be enabled,
by hand, case by case. No checking is done so the user must take
responsibility for ensuring that only syntactically and semantically
valid options are passed to the relevant principal command.

Signed-off-by: Phill Kelley <34226495+Paraphraser@users.noreply.github.com>