add support for LOCAL backup/restore
The `iotstack_backup` and `iotstack_restore` methods have assumed one
of SCP, RSYNC or RCLONE has been configured prior to the run, such that
the backup can save the backup files off the local host, and the restore
can fetch the previously-saved backup files across the network.

This change adds support for a LOCAL method which only saves backup
files in the local `~/IOTstack/backups` directory; and only searches
that directory on restore.

Although it is not particularly safe to keep both the working versions
(those in `~/IOTstack/volumes`) and backups on the same machine, the
LOCAL method can be useful for someone who wants to use IOTstackBackup
as part of a rebuild process - just get a local backup on the existing
machine, copy the files using some sensible method, move the files into
place on the rebuilt machine, run the restore.

And then, hopefully, think about configuring IOTstackBackup properly...

Signed-off-by: Phill Kelley <34226495+Paraphraser@users.noreply.github.com>