add migration assistant
Adds a pair of scripts to support migration. Intended for use where
the OLD system has NOT been running IOTstackBackup while the NEW
system was build with PiBuilder (which installs IOTstackBackup).

In essence, `iotstack_migration_backup` calls the same subordinate
scripts as `iotstack_backup` but writes the results into the current
working directory with a runtag of "migration".

Similarly, `iotstack_migration_restore` calls the same subordinate
scripts as `iotstack_restore` but expects to find the restore sources
(the .tar, .tar.gz, etc) in the current working directory.

This approach follows on from a discussion with Andreas Spiess about
migrating from a Raspberry Pi to a Proxmox-based Debian guest.

Signed-off-by: Phill Kelley <34226495+Paraphraser@users.noreply.github.com>