exclude motioneye video captures from backups
Depending on configuration choices like frame rate, the ±time around a
motion trigger that you wish to preserve, and the numbers of days of
motion history you wish to preserve, the folder containing this video
can be **massive**. This change omits it by default.

Also elaborates on the reasons for other default exclusions.

Signed-off-by: Phill Kelley <34226495+Paraphraser@users.noreply.github.com>