fixes isStackUp() function to not fail when foreign containers are running
An exchange on Discord beginning at
[this post](https://discord.com/channels/638610460567928832/638610461109256194/1204978271431819274)
led to the discovery of a bug in the `isStackUp()` function.

The bug manifested when a container was running that had been started
via `docker run` (ie external to IOTstack).

The previous version of `isStackUp()` used `jq` to filter the
`com.docker.compose.project` lines that matched the project
(ie "iotstack" - all lower case). The problem is that containers
that are not part of any project return NULL so the original `jq`
`contains(\"$1\")"` filter will complain saying:

```
jq: error (at <stdin>:1): null (null) and string ("iotstack") cannot have their containment checked
```

The previous version of `isStackUp()` piped the filtered output to
`wc` for counting the hits. The revised version doesn't use `jq` to
do the filtering, instead using `grep` to do both the filtering and
counting.

Signed-off-by: Phill Kelley <34226495+Paraphraser@users.noreply.github.com>