automatic support for backup and restore of continuous queries
Previously, the "epilog" file was the only mechanism for restoring user
credentials, grants and continuous queries.

With this update, the "epilog" is only needed for user credentials and
grants. The former can't be automated (because passwords can't be
recovered), and the latter can't reasonably be automated without the
former.

Continuous queries are now extracted during `iotstack_backup_influxdb`
and saved as `continuous-queries.influxql` (a plain text file). That
file becomes part of the `.tar`.

If `continuous-queries.influxql` is present after the `.tar` is
unpacked, `iotstack_restore_influxdb` will pass the file to `influx`
to recreate the continuous queries.

Note that reloading of duplicate CQs via the epilog mechanism has no
effect. However, because the epilog is loaded *after* the CQs, any
differences in definition will be resolved in favour of the epilog.

Signed-off-by: Phill Kelley <34226495+Paraphraser@users.noreply.github.com>