mysql to mariadb command name changes
Commands like `mysql`, `mysqladmin` and `mysqldump` are being deprecated
in favour of `mariadb`, `mariadb-admin` and `mariadb-dump`. Affected
scripts now check for the availability of the newer commands and use
those, falling back to the older commands where necessary.

Also, due to an apparent bug in the linuxserver implementation, the
root password is not being set on newly-initialised databases. The
restore scripts try to deal with this problem by assuming the root
password has not been set and set it explicitly to whatever value
is passed into the container via `MYSQL_ROOT_PASSWORD`.

See [Issue 163](https://github.com/linuxserver/docker-mariadb/issues/163)

Signed-off-by: Phill Kelley <34226495+Paraphraser@users.noreply.github.com>