alternative approach to root password situations for mariadb instances
When I filed https://github.com/linuxserver/docker-mariadb/issues/163 I
assumed this was a bug but it turned out to be a feature needed to
support logrotate. That means any action by either IOTstack or
IOTstackBackup to forcibly apply the root password from within the
container will defeat that feature.

I have closed my IOTstack PRs 804 and 805.

MariaDB instances can be in the following situations:

1. No root password defined via environment var but mariadb commands
   expect a password a password inside the container.

2. No root password defined via environment var but mariadb commands do
   not accept a password anyway.

3. Root password defined but mariadb commands do not accept a password.

4. Root password defined and mariadb commands expect a root password.

IOTstackBackup backups and restores will succeed in cases 2, 3 and 4.

Case 1 is a user error which will need to be fixed by defining
`MYSQL_ROOT_PASSWORD`. Unfortunately, backups (almost certainly) and
restores (possibly) will fail, silently, with the only evidence being
in the log.

Signed-off-by: Phill Kelley <34226495+Paraphraser@users.noreply.github.com>